ajout d'un séance depuis le front jusq'a la base de donnée fonctionelle

+++ src/main/java/Pecadillo/isika/al/dao/SeanceDao.java
@@ -8,5 +8,5 @@ import Pecadillo.isika.al.model.Seance;
 
 
 @Repository
-public interface SeanceDao extends JpaRepository<Seance, Integer> {
+public interface SeanceDao extends JpaRepository<Seance, Long> {
 }
